Moved all variables into a global location. This allows for better communication around the application.
Updated the user login script to not cause a crash when closing the scanner screen. Also updated the icons for the user login button to change when logged in or out.

Added the buttons for the iOS side of the application. This won't work on Android, more work is needed before the user can add, delete or edit devices.

// To Do \\

1. Add the buttons to allow a user to add, remove or edit devices.
2. Find a way to close the scanner while the screen is not focused.
3. Neaten up the code and remove uncessary API calls when linking devices.